requirements.txt
https://www.djangoproject.com/download/1.7c1/tarball/
dj-database-url
DjangoRestless
psycopg2
